+++ test/chorus.ml
@@ -22,7 +22,10 @@ let chorus_t = Term.(const chorus $ count $ msg)
 
 let info =
   let doc = "print a customizable message repeatedly" in
-  let man = [ `S "BUGS"; `P "Email bug reports to <hehey at example.org>.";] in
+  let man = [
+    `S Manpage.s_bugs;
+    `P "Email bug reports to <hehey at example.org>."]
+  in
   Term.info "chorus" ~version:"1.6.1" ~doc ~man
 
-let () = match Term.eval (chorus_t, info)  with `Error _ -> exit 1 | _ -> exit 0
+let () = match Term.eval (chorus_t, info) with `Error _ -> exit 1 | _ -> exit 0